requirements.txt
pandas
numpy
matplotlib
polygon-api-client
